requirements.txt
streamlit
langchain
langchain-community
langchain-openai
faiss-cpu
sentence-transformers
pypdf
python-dotenv
tiktoken